+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/11)
 * Install react-router-dom and implement router
+* Add Metrics and Selector nav items routing
 
 0.20 (2024-06-25)
 ===============